JH-change LoadingPage from a page into a component
