+++ src/App.js
@@ -2,6 +2,7 @@ import { Routes, Route } from "react-router-dom";
 import Home from "./main.js";
 import PromptPage from "./pages/PromptPage/PromptPage";
 import TestPage from "./pages/TestPage/TestPage";
+import LoadingPage from "./pages/LoadingPage/LoadingPage";
 
 function App() {
   return (
@@ -10,6 +11,7 @@ function App() {
         <Route path="/" element={<Home />} />
         <Route path="/prompt" element={<PromptPage />} />
         <Route path="/test" element={<TestPage />} />
+        <Route path="/loading" element={<LoadingPage />} />
       </Routes>
     </>
   );
