+++ src/components/LoadingPage/LoadingPage.scss
@@ -3,30 +3,18 @@
 @import "../../styles/partials/typography.scss";
 
 .loading-page {
-  background-color: $Powder-Blue;
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  min-height: 100vh;
+  background-color: $MVP-Yellow;
+  border-radius: 20px;
+  border: 4px solid #000;
+  box-sizing: border-box;
+  width: 70%;
+  height: 64.1vh;
   position: absolute;
-  top: 0;
-  z-index: -1;
-
-  &__container {
-    background-color: $MVP-Yellow;
-    border-radius: 20px;
-    border: 4px solid #000;
-    box-sizing: border-box;
-    width: 70%;
-    height: 64.1vh;
-    position: absolute;
-    top: 9.5514rem;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-  }
+  top: 9.5514rem;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
 
   &__icon {
     margin-top: 2.94rem;
