+++ src/pages/PromptPage/PromptPage.jsx
@@ -1,5 +1,6 @@
 import "./PromptPage.scss";
 import DashboardNavbar from "../../components/DashboardNavbar/DashboardNavbar";
+import LoadingPage from "../../components/LoadingPage/LoadingPage";
 
 const PromptPage = () => {
   return (
@@ -7,11 +8,13 @@ const PromptPage = () => {
       <div className="promptpage__navbar">
         <DashboardNavbar />
       </div>
-
       <div className="promptpage">
         <div className="promptpage__container">
           <div className="promptpage__top-bar" />
-          <div className="promptpage__middle-container">Place some Content here...</div>
+          <div className="promptpage__middle-container">
+            Place some Content here...
+            <LoadingPage />
+          </div>
           <div className="promptpage__bottom-bar" />
         </div>
       </div>
